+++ flappy/src/flappyEntities/logic/LogicComponent.java
@@ -1,5 +1,5 @@
 package flappyEntities.logic;
 
 public interface LogicComponent {
-    void update(int i);
+    void update(double i);
 }
